+++ apps/meteor/app/reactions/client/methods/setReaction.ts
@@ -1,4 +1,4 @@
-import type { IMessage, IRoom } from '@rocket.chat/core-typings';
+import type { IMessage } from '@rocket.chat/core-typings';
 import type { ServerMethods } from '@rocket.chat/ddp-client';
 import { Meteor } from 'meteor/meteor';
 
@@ -23,7 +23,7 @@ Meteor.methods<ServerMethods>({
 			return false;
 		}
 
-		const room: IRoom | undefined = Rooms.findOne({ _id: message.rid });
+		const room = Rooms.state.get(message.rid);
 		if (!room) {
 			return false;
 		}
@@ -36,11 +36,11 @@ Meteor.methods<ServerMethods>({
 			return false;
 		}
 
-		if (roomCoordinator.readOnly(room._id, user)) {
+		if (roomCoordinator.readOnly(room, user)) {
 			return false;
 		}
 
-		if (!Subscriptions.findOne({ rid: message.rid })) {
+		if (!Subscriptions.state.find(({ rid }) => rid === message.rid)) {
 			return false;
 		}