+++ apps/meteor/ee/app/livechat-enterprise/server/index.ts
@@ -1,4 +1,5 @@
 import { License } from '@rocket.chat/license';
+import { patchOmniCore } from '@rocket.chat/omni-core-ee';
 import { Meteor } from 'meteor/meteor';
 
 import './methods/addMonitor';
@@ -29,6 +30,8 @@ import './lib/AutoCloseOnHoldScheduler';
 import './business-hour';
 import './api';
 import { createDefaultPriorities } from './priorities';
+
+patchOmniCore();
 
 await License.onLicense('livechat-enterprise', async () => {
 	require('./hooks');