+++ apps/meteor/tests/e2e/page-objects/login.ts
@@ -10,7 +10,7 @@ export class LoginPage {
 	constructor(protected readonly page: Page) {}
 
 	get loginButton() {
-		return this.page.getByRole('button', { name: 'Login' });
+		return this.page.getByRole('button', { name: 'Login', exact: true });
 	}
 
 	/** @deprecated ideally the previous action should ensure the user is logged out and we should just assume to be at the login page */