+++ packages/rest-typings/src/v1/chat.ts
@@ -175,24 +175,6 @@ const ChatUnstarMessageSchema = {
 
 export const isChatUnstarMessageProps = ajv.compile<ChatUnstarMessage>(ChatUnstarMessageSchema);
 
-type ChatPinMessage = {
-	messageId: IMessage['_id'];
-};
-
-const ChatPinMessageSchema = {
-	type: 'object',
-	properties: {
-		messageId: {
-			type: 'string',
-			minLength: 1,
-		},
-	},
-	required: ['messageId'],
-	additionalProperties: false,
-};
-
-export const isChatPinMessageProps = ajv.compile<ChatPinMessage>(ChatPinMessageSchema);
-
 type ChatUnpinMessage = {
 	messageId: IMessage['_id'];
 };
@@ -830,6 +812,9 @@ const ChatPostMessageSchema = {
 					type: 'object',
 					nullable: true,
 				},
+				parseUrls: {
+					type: 'boolean',
+				},
 			},
 			required: ['roomId'],
 			additionalProperties: false,
@@ -874,6 +859,9 @@ const ChatPostMessageSchema = {
 				customFields: {
 					type: 'object',
 					nullable: true,
+				},
+				parseUrls: {
+					type: 'boolean',
 				},
 			},
 			required: ['channel'],
@@ -945,11 +933,6 @@ export type ChatEndpoints = {
 	};
 	'/v1/chat.unStarMessage': {
 		POST: (params: ChatUnstarMessage) => void;
-	};
-	'/v1/chat.pinMessage': {
-		POST: (params: ChatPinMessage) => {
-			message: IMessage;
-		};
 	};
 	'/v1/chat.unPinMessage': {
 		POST: (params: ChatUnpinMessage) => void;