+++ packages/ui-contexts/src/index.ts
@@ -95,9 +95,10 @@ export { useSetInputMediaDevice } from './hooks/useSetInputMediaDevice';
 export { useAccountsCustomFields } from './hooks/useAccountsCustomFields';
 export { useUserPresence } from './hooks/useUserPresence';
 export { useUnstoreLoginToken } from './hooks/useUnstoreLoginToken';
+export { useOnLogout } from './hooks/useOnLogout';
 
 export { UploadResult } from './ServerContext';
 export { TranslationKey, TranslationLanguage } from './TranslationContext';
-export { Fields } from './UserContext';
+export { Fields, FindOptions } from './UserContext';
 
 export { SubscriptionWithRoom } from './types/SubscriptionWithRoom';